+++ lib/vanagon/cli/render.rb
@@ -1,5 +1,6 @@
 require 'docopt'
 require 'json'
+require 'vanagon/logger'
 
 class Vanagon
   class CLI
@@ -27,7 +28,7 @@ class Vanagon
       def parse(argv)
         Docopt.docopt(DOCUMENTATION, { argv: argv })
       rescue Docopt::Exit => e
-        puts e.message
+        VanagonLogger.error e.message
         exit 1
       end